fix some bugs, make holograms work

+++ src/main/java/lol/pyr/znpcsplus/metadata/V1_8Factory.java
@@ -16,7 +16,7 @@ public class V1_8Factory implements MetadataFactory {
 
     @Override
     public EntityData effects(boolean onFire, boolean glowing, boolean invisible) {
-        return new EntityData(0, EntityDataTypes.BYTE, (onFire ? 0x01 : 0) | (invisible ? 0x20 : 0));
+        return new EntityData(0, EntityDataTypes.BYTE, (byte) ((onFire ? 0x01 : 0) | (invisible ? 0x20 : 0)));
     }
 
     @Override
